backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
scipy==1.14.0
numpy==1.26.4
firebase-admin==6.5.0
google-generativeai==0.8.0
pydantic==2.9.0
pydantic-settings==2.5.0
httpx==0.27.0
python-dotenv==1.0.0
